F16 - soil profile calculates moisture based on distance and can be user input

+++ island/src/main/java/ca/mcmaster/cas/se2aa4/a3/island/Humidity/DrySoil.java
@@ -3,7 +3,7 @@ package ca.mcmaster.cas.se2aa4.a3.island.Humidity;
 public class DrySoil implements SoilProfile {
 
     @Override
-    public int calcMoisture(int moisture, int distance) {
+    public int calcMoisture(int moisture, double distance) {
         return (int)(moisture * (30 / Math.sqrt(distance)));
     }
 }
